Renew config and wsw data packet.
Renew config and wsw data packet.

+++ WswPlatform/DataPacketToWsw.cs
@@ -5,8 +5,8 @@ namespace TeachingPlatformApp.WswPlatform
     [StructLayout(LayoutKind.Sequential, Pack = 1)]
     public struct DataPacketToWsw
     {
-        public byte Header { get; set; }
-        public byte Index { get; set; }
+        public double Header { get; set; }
+        public double Index { get; set; }
         public double X { get; set; }
         public double Y { get; set; }
     }
